KafkaBackend/requirements.txt
Flask==3.1.0 
Flask-Cors==5.0.0
Jinja2==3.1.5 
MarkupSafe==3.0.2 
Werkzeug==3.1.3 
blinker==1.9.0 
cassandra-driver==3.29.2 
certifi==2024.12.14 
charset-normalizer==3.4.1 
click==8.1.8 
colorama==0.4.6 
confluent-kafka==2.7.0 
geomet==0.2.1.post1 
idna==3.10 
itsdangerous==2.2.0 
jsonify==0.5 
pip==23.0.1
requests==2.32.3 
setuptools==65.5.0 
six==1.17.0 
urlib3==2.3.0 